requirements.txt
django-tailwind[reload]
